+++ src/types/ContactFormData.ts
@@ -1,0 +1,9 @@
+export type ContactFormData = {
+  firstName: string;
+  lastName: string;
+  telephoneNumber: string;
+  company?: string;
+  email: string;
+  subject?: string;
+  message: string;
+};