+++ src/utils/index.ts
@@ -1,3 +1,5 @@
 export { toKebabCase } from "./toKebabCase";
 export { getCurrentYear } from "./getCurrentYear";
 export { classNames } from "./classNames";
+export { isValidEmail } from "./isValidEmail";
+export { sendEmail } from "./sendEmail";